refactor(utils): change getDiffInCalendarMonths
